Allow for a multiple dbs per arango cluster
Curently if several DBs used on a same arango cluster it would not apply
migrations due to the fact that migrations are stored in a single MASTER
index of migrations, while it would be cool to have support for per DB
configuration where migrations would be applied if different DB is used.

Example might be the case when there're several envs used on a same
cluster, aka production, stage, nightly etc.